Fixed tests for 1.9.2. Added utility script for use w/ rvm.

+++ ext/gmpbench_timing.c
@@ -41,7 +41,7 @@ cputime ()
   struct rusage rus;
 
   getrusage (0, &rus);
-  return rus.ru_utime.tv_sec * 1000 + rus.ru_utime.tv_usec / 1000;
+  return (int) (rus.ru_utime.tv_sec * 1000 + rus.ru_utime.tv_usec / 1000);
 }
 #endif
